Update AboutMe section and components
