+++ src/components/AboutMe/index.js
@@ -1,0 +1,3 @@
+import AboutMe from './AboutMe';
+
+export default AboutMe;
